Clean up SdbData
